refactor: `Memo~` DTO 명칭 변경(`CheckInMemo~`) 및 `PresignedUrlItem` 기존 모델 사용하도록 수정
