+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/data/dto/request/checkin/CheckInMemoRequest.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class MemoRequest(
+data class CheckInMemoRequest(
     @SerialName("memo")
     val content: String,
 )
