+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/data/dto/response/checkin/CheckInMemoResponse.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class MemoResponse(
+data class CheckInMemoResponse(
     @SerialName("memo")
     val memo: String?, // 메모 내용 (없으면 null)
 )
